Refactor: Extract .gen_collector_thread
