+++ lib/slow_enumerator_tools.rb
@@ -15,6 +15,9 @@ module SlowEnumeratorTools
 end
 
 require_relative 'slow_enumerator_tools/version'
+
+require_relative 'slow_enumerator_tools/util'
+
 require_relative 'slow_enumerator_tools/batcher'
 require_relative 'slow_enumerator_tools/bufferer'
 require_relative 'slow_enumerator_tools/merger'
