+++ lib/slow_enumerator_tools/bufferer.rb
@@ -2,25 +2,10 @@
 
 module SlowEnumeratorTools
   module Bufferer
-    STOP_OK = Object.new
-    STOP_ERR = Object.new
-
     def self.buffer(enum, size)
       queue = SizedQueue.new(size)
-      thread = gen_collector_thread(enum, queue)
+      thread = SlowEnumeratorTools::Util.gen_collector_thread(enum, queue)
       gen_enumerator(queue, thread)
-    end
-
-    def self.gen_collector_thread(enum, queue)
-      Thread.new do
-        begin
-          enum.each { |e| queue << e }
-          queue << STOP_OK
-        rescue StandardError => e
-          queue << STOP_ERR
-          queue << e
-        end
-      end
     end
 
     def self.gen_enumerator(queue, collector_thread)
@@ -28,9 +13,9 @@ module SlowEnumeratorTools
         loop do
           e = queue.pop
 
-          if STOP_OK.equal?(e)
+          if SlowEnumeratorTools::Util::STOP_OK.equal?(e)
             break
-          elsif STOP_ERR.equal?(e)
+          elsif SlowEnumeratorTools::Util::STOP_ERR.equal?(e)
             raise queue.pop
           end
